+++ spec/lib/govuk_publishing_components/presenters/shared_helper_spec.rb
@@ -1,43 +1,43 @@
 RSpec.describe GovukPublishingComponents::Presenters::SharedHelper do
   describe "Shared component helper" do
-    after(:each) do
+    after do
       I18n.locale = :en
     end
 
     it "returns a default heading level" do
-      shared_helper = GovukPublishingComponents::Presenters::SharedHelper.new({})
+      shared_helper = described_class.new({})
       heading = shared_helper.get_heading_level
       expect(heading).to eql("h2")
     end
 
     it "returns a given heading level" do
-      shared_helper = GovukPublishingComponents::Presenters::SharedHelper.new(heading_level: 6)
+      shared_helper = described_class.new(heading_level: 6)
       heading = shared_helper.get_heading_level
       expect(heading).to eql("h6")
     end
 
     it "returns the default heading level if passed value is wrong" do
-      shared_helper = GovukPublishingComponents::Presenters::SharedHelper.new(heading_level: 9)
+      shared_helper = described_class.new(heading_level: 9)
       heading = shared_helper.get_heading_level
       expect(heading).to eql("h2")
     end
 
     it "returns a span instead of a heading if heading level is 0" do
-      shared_helper = GovukPublishingComponents::Presenters::SharedHelper.new(heading_level: 0)
+      shared_helper = described_class.new(heading_level: 0)
       result = shared_helper.get_heading_level
       expect(result).to eql("span")
     end
 
     it "returns nil if given locale is same as page locale" do
       default_locale = I18n.locale
-      shared_helper = GovukPublishingComponents::Presenters::SharedHelper.new({})
+      shared_helper = described_class.new({})
 
-      expect(shared_helper.t_locale_check(default_locale)).to be nil
+      expect(shared_helper.t_locale_check(default_locale)).to be_nil
     end
 
     it "returns a locale if different to the page locale" do
       locale = "ar"
-      shared_helper = GovukPublishingComponents::Presenters::SharedHelper.new({})
+      shared_helper = described_class.new({})
 
       expect(shared_helper.t_locale_check(locale)).to eq locale
     end
@@ -46,7 +46,7 @@ RSpec.describe GovukPublishingComponents::Presenters::SharedHelper do
       I18n.locale = :de
 
       translation_key = "this.is.a.key.that.should.not.be.found.to.test.the.translation"
-      shared_helper = GovukPublishingComponents::Presenters::SharedHelper.new({})
+      shared_helper = described_class.new({})
 
       expect(shared_helper.t_lang(translation_key)).to eq "lang=en"
     end
@@ -55,16 +55,16 @@ RSpec.describe GovukPublishingComponents::Presenters::SharedHelper do
       I18n.locale = :fr
 
       translation_key = "components.contents_list.contents"
-      shared_helper = GovukPublishingComponents::Presenters::SharedHelper.new({})
+      shared_helper = described_class.new({})
 
-      expect(shared_helper.t_lang(translation_key)).to be nil
+      expect(shared_helper.t_lang(translation_key)).to be_nil
     end
 
     it "returns the locale if translation is not present" do
       I18n.locale = :de
 
       translation_key = "this.is.a.key.that.should.not.be.found.to.test.the.translation"
-      shared_helper = GovukPublishingComponents::Presenters::SharedHelper.new({})
+      shared_helper = described_class.new({})
 
       expect(shared_helper.t_locale(translation_key)).to eq :en
     end
@@ -73,7 +73,7 @@ RSpec.describe GovukPublishingComponents::Presenters::SharedHelper do
       I18n.locale = :fr
 
       translation_key = "components.contents_list.contents"
-      shared_helper = GovukPublishingComponents::Presenters::SharedHelper.new({})
+      shared_helper = described_class.new({})
 
       expect(shared_helper.t_locale(translation_key)).to be :fr
     end